Add files via upload
As we know that mental health is an important thing for a person's life, it plays a fundamental role in his social and family life and everything, and it is a fundamental factor that makes life stable. In this topic, I focus on analyzing and deducing information about people who are exposed to mental illnesses and people who have already had previous tragic experiences. Through analyzing a group of factors that may be a reason for a person to be exposed to the disease, I analyze this data based on algebra, statistics and mathematics using a group of tools Pandas and Numpy And Seaborn You Can Discover This By Your Self in My Notebook